Create tables
The purpose of the tables herein are self-explicative
except perhaps, scripts and itsps.

scripts: This table will be essential for sending channels
to the right context based on the script (automated attendant)
a user wants to use for their calls. The value in the id column
for a given record will be combined with the corresponding user_id
to identify a context in the dialplan. The content column simply
stores, for reference, the dialplan code within the identified
context.

itsps: This stores the name that is used to identify an endpoint
(SIP trunk) in the PBX. It will be useful for dyamically selecting
a trunk for call termination.

dids: This stores the phone numbers that can be used as a valid
caller id for calls terminated from a SIP trunk. The DID and SIP
trunk must be provided by the same ITSP, hence, the itsp_id
foreign key.