Declare start
This is the controller used to initiate outbound
dialing. It will read the phone numbers from the
specified sheet then invoke the originate function
for each number starting from the specified index.

The number of calls created is decided by the
CONCURRENT_CALLS environment variable.
The endpoint (SIP Trunk) used to terminate the
call is dependent on the caller id the user
selected.

Note: createContext uses the user's id and the
specified script id to generate dialplan context
that should be used for the call.