Cache campaign resources and set dialer status
cacheCampaignResources is invoked to cache the phone numbers
and id of the script selected for a campaign. This avoids the
for database reads as the campaign progresses.

The dialer status (set as an env variable) is used to prevent
user's from having multiple ongoing campaigns and allow them
to pause, resume and stop the dialer.