Declare Queue
An instance of this class will be used to implement a
queue. The encapsulated private properties and methods
are used to manage internal state, while the instance methods
are used to push and pop calls to/from the queue.

1. #setRoundRobinPointer - This method moves the pointer (index)
used to select the endpoint that should be used for a call.
Invoking this mehtod within 'queue' method ensures calls are evenly
distrubuted across the available endpoints.

2. #endpount_output_count - This property tracks endpoint usage
(how many times each endpoint is used). The propety is updated by
#incrementOutputCount which is invoked within the queue method.

3. #removeEndpoint - This method is used to remove an endpoint
from the list of endpoints that can be used. It is invoked within the
queue method when the endpoint's usage reaches
#max_output_per_endpoint. Note, the method also resets the
#round_robin_pointer to ensure calls are evenly distributed across
the remaining endpoints.