Declare getScripts and getScriptAudio
getScripts and getScriptAudio fetches all of a user's scripts
and the audio BLOB associated with a specified script, respectively.

Also, added a field called ‘audio’ to the scripts table for storing the audio blob
(the greeting only) for the the script, and add NOT NULL constraint to the
content column.