Modify start
Rather than using a single endpoint and caller id
for a campaign, the application will use the round
robin method to evenly distribute calls among all
endpoints and the user's dids (caller ids).

Hence, a caller_id is no longer a required request
paylod parameter, the controller doesn't need to
fetch an endpoint, and a different caller id is passed
to the queue_call event emission on every iteration
of the for loop.