Declare setCampaignIndicies
The function caches the following indicies:

- The first number called
- The next number to be called.
- Parameters used to determine the next caller id to use

Note, the length of the user's caller id list is used
to calculate the actual index of the next caller id
to use. The value in the next_caller_id field is merely
total number of calls made plus one. To determine the
caller_id to use for the next call the application will
use simply calculate next_caller_id modulo count_caller_id.