Modify setCampaignIndicies
The start_index field should only be set once
throughout a campaign. Passing null or undefined
as the value of the start_index param indicates
that the start_index field should not be set, thus
, null or undefined should be passed as the second
argument to this function whenever a campaign is
resumed or progressing to the next call in line.