Modify getOngoingCampaignResources
If a value is passed for the args argument, the function
will fetch the script id, a subset of calling list, and
all the caller ids for the user's ongoing campaign.

This version of the function will be used when resuming
a campaign or progressing to the next call in line.