Modify toggle
When the user resumes a paused campaign, the contoller
invokes getDetailsToContinueCampaign to retrieve then
next n phone numbers to be called, the user's caller_ids,
the id of the script being used for the campaign, and the
indicies of the next number to call and the caller id to
use.

These values are used queue as many calls as the concurrency
env variable. Afterwards, setCampaignIndicies is invoked to
update the campaign's progress accordingly.