Use a redis hash to store call stauses
Pushing calling statuses to a list makes a fundamentally
erroneous assumption - that call statuses are received
in chronological order.

Since call statuses are received in no praticular order,
it's only correct to use a hash which associates the
index of the number called with the status of the call.

Also, I removed 'invalid number' from the regEx patter that
determines if a status should be cached or not.