Declare createChannelVariables
This function returns a JSON object that will be used as
the body in requests sent to the POST /channels ARI
endpoint. The objects key/value pairs will be used to
create channel variables on the channel created.

In this case, it's used to create a channel variable
for each agent's phone number and another variable
called i which is simply a count of how many
variables the object contains.

Also, I exported formatNumbers because it's used
inside createChannelVariables.