Get agents' phone number and cache channel variables
When a user starts or resumes a campaign, the respective controller
will fetch all their agents' phone number from the DB, use the
phone numbers to create a channel variable JSON object then cache
the object.

The JSON object will be fetched from cache and sent in the request
body whenever a channel is being created for this user. This
is how Asterisk will know the phone numbers of the user's agents
should it need to dial them.