Define channel variables
The channel_variables param on the originate function
will be used to define channel variables in the body
of requests to the POST /channels ARI endpoint.

In this case, it is used to pass the phone numbers
of agents to the ARI so the dialplan may know the
numbers to dial should it need to.